Siga-Sr: exibindo contagem de tempo mesmo sem haver prazo

+++ sigasr/app/models/SrAcordo.java
@@ -147,7 +147,7 @@ public class SrAcordo extends HistoricoSuporte implements Selecionavel {
 		return getAtributo(att);
 	}
 
-	public Integer getAtributoEmSegundos(String codigo) {
+	public Long getAtributoEmSegundos(String codigo) {
 		SrAtributoAcordo pa = getAtributo(codigo);
 		if (pa == null)
 			return null;
