+++ sigasr/app/models/SrAtributoAcordo.java
@@ -38,7 +38,7 @@ public class SrAtributoAcordo extends HistoricoSuporte {
 	public SrOperador operador;
 
 	@Column(name = "VALOR")
-	public Integer valor;
+	public Long valor;
 
 	@ManyToOne
 	@JoinColumn(name = "UNIDADE_MEDIDA")
@@ -102,7 +102,7 @@ public class SrAtributoAcordo extends HistoricoSuporte {
 				.getIdUnidadeMedida().intValue() : null);
 	}
 	
-	public Integer getValorEmSegundos(){
+	public Long getValorEmSegundos(){
 		return getValorEUnidade().getValorEmSegundos();
 	}
